✨ feat: shared topic page document title + Copy Link menu action
- Document title now shows "<topic> · LobeHub": relocate the share topic
  route meta to the route folder and wire it into both desktop router
  configs (it was previously only registered on mobile, so desktop fell
  back to just "LobeHub")
- Add a "Copy Link" item to the share page header menu that copies the
  current page URL to the clipboard
