+++ src/routes/share/t/[id]/_layout/HeaderMenu.tsx
@@ -1,9 +1,10 @@
 'use client';
 
-import { ActionIcon } from '@lobehub/ui';
+import { ActionIcon, copyToClipboard } from '@lobehub/ui';
 import { type DropdownItem, DropdownMenu } from '@lobehub/ui/base-ui';
-import { ExternalLink, Flag, MoreHorizontal } from 'lucide-react';
-import { memo, useMemo } from 'react';
+import { App } from 'antd';
+import { ExternalLink, Flag, LinkIcon, MoreHorizontal } from 'lucide-react';
+import { memo, useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { mailTo, OFFICIAL_SITE } from '@/const/url';
@@ -12,9 +13,21 @@ const REPORT_EMAIL = 'hi@lobehub.com';
 
 const HeaderMenu = memo(() => {
   const { t } = useTranslation('chat');
+  const { message } = App.useApp();
+
+  const handleCopyLink = useCallback(async () => {
+    await copyToClipboard(window.location.href);
+    message.success(t('shareModal.copyLinkSuccess'));
+  }, [message, t]);
 
   const items = useMemo<DropdownItem[]>(
     () => [
+      {
+        icon: <LinkIcon size={16} />,
+        key: 'copy-link',
+        label: t('sharePage.menu.copyLink'),
+        onClick: handleCopyLink,
+      },
       {
         icon: <ExternalLink size={16} />,
         key: 'go-to-lobehub',
@@ -30,7 +43,7 @@ const HeaderMenu = memo(() => {
         label: <a href={mailTo(REPORT_EMAIL)}>{t('sharePage.menu.report')}</a>,
       },
     ],
-    [t],
+    [t, handleCopyLink],
   );
 
   return (
